rename components and utils directory namespace to add _ and fix some import method
