+++ src/app/_components/SlideAnimation.tsx
