+++ src/app/_utils/utility_function.ts
