+++ src/app/statistics/page.tsx
@@ -13,7 +13,7 @@ import { MonthProvider, MonthContext } from '@components/MonthPaginator'
 import MonthPaginator from '@components/MonthPaginator'
 
 import { APIClient } from '@utils/api_client'
-import { HouseholdMonthlySummaryResponse } from "@/app/utils/interfaces"
+import { HouseholdMonthlySummaryResponse } from "@utils/interfaces"
 
 
 const client = new APIClient()
