+++ src/app/task/detail/page.tsx
@@ -10,13 +10,13 @@ import DOMPurify from 'dompurify'
 
 import { YearContext } from '@components/YearPicker'
 import { MonthContext } from '@components/MonthPaginator'
-import { TextLinkToBackToPreviousPage } from '@/app/components/TextLink'
-import { EditButton, DeleteButton } from '@/app/components/Buttons'
+import { TextLinkToBackToPreviousPage } from '@components/TextLink'
+import { EditButton, DeleteButton } from '@components/Buttons'
 import { ToasterComponent, APIResponseToast } from '@components/ToastMessage'
 import { CreatedBy } from '@utils/constants'
 
 import { APIClient } from '@utils/api_client'
-import { TaskData, TaskResponse, TaskCommentData, TaskCommentResponse } from '@/app/utils/interfaces'
+import { TaskData, TaskResponse, TaskCommentData, TaskCommentResponse } from '@utils/interfaces'
 import { convertUrlsToLinks, setCreatedByStr, setStatusStr, setPriorityStr, splitYearMonthDayStr,
          getToday, getDate, getNumberOfDays, getWeekDay, getMonthArray, getCurrentDateTime } from '@utils/utility_function'
 import { ReactQuillStyles } from '@utils/styles'
